Solve the arithmetic problem with parenthesis
The arithmetic operator had a problem when trying to parse something
such as :

`myVar = (a + b) * c`

It appears that for parser to work, we need to see the parenthesis as
a term in itself composed of a series of terms. This creates a
perfectly valid recursive-loops where a arithmetic expression can
become a term, and a single term could be seen as an arithmetic
expression.

Nevertheless, we might change the parser and get inspiration from the
C yacc files in which the precedence is devided differently and
considers all the operators (including logical, comparison, arithmetic,
shifting).